+++ CHANGES.md
@@ -2,6 +2,7 @@ v0.2.12 (dev)
 =============
 + add checkbox to HTML to scale IBS0, IBS2, etc by number of sites shared by the samples. this almost always
   results in a scaling that is better across (pairs of) samples.
++ ancestry: allow globs for ancestry files (#59)
 
 
 v0.2.11